We focused on diagnosing faults in rotating machinery.
Extreme conditions and frequent changes in workload pose significant risks for rotary machinery widely used in industrial production. When rotary machinery fails unexpectedly, it can lead to disastrous consequences. In order to ensure the continued functionality and security of machinery, a method for fault diagnosis must be developed for use with rotary machinery